Number entry part: wiggle room is at the 12th sig fig
fixes #919

Previously, when the min or max value for a number entry part was set
using a `number` type value, wiggle room of ±1e-12 was added to account
for floating-point error.

When the number is large, this difference can't be represented, so
there's effectively no wiggle room. So when there's a floating-point
error in the calculation, it looks like the correct answer is not
accepted.

This changes the wiggle room to take place at around the 12th
significant figure. If the value is on the order of 10^n, the wiggle
room is 10^(n-12).