throw an error when an HTML value is already in the page.
In DOMcontentsubber, when inserting an HTML element, check that it
doesn't already have a parent. If it does, then it's being inserted for
a second time, which is a common error.

This throws an error "An HTML value has been embedded twice. Consider
defining a function to generate a new value each time it is used."

fixes #886