allow more characters in variable names
fixes #787

Variable names may now start with a `$`, a `_`, or any of the character
classes: Lu, Ll, Lo, Lt. The rest may be made of `_`, or any of the
character classes Lu, Ll, Lo, Lt, Nl, Nd.

Names are not run through Unicode normalisation, so diacritics applied
with modifier characters are not equivalent to single-character letters
with diacritics.
I want to look at that more carefully before doing anything.