re_name allows curly braces around names
I removed this, thinking it wasn't needed, but it turns out that it was
the last bit of code allowing people to use surplus curly braces in JME
expressions!

This commit is a quick fix to get things working again. I'd like to move
this to the punctuation token, so that curly braces are silently
interpreted as parentheses.