better logic for detecting duplicate HTML embeds
My first go for a fix for #886, in
94b3648c8dd0d3aa9cbb468204d88de7b0c10850, tried to detect duplicate
embeds by checking if the embedded element already had a parent.

The old JS method for creating JSXGraph diaagrams attaches the board to
an invisible parent element in the page, triggering this check.

This commit changes the logic to add a `numbas_embedded` property to the
element being embedded. The error is thrown if an element with this
property already set tries to be embedded again.