All* events and signals covered by tests and documentation
There are unit tests checking each of the signals and events triggered
by exams, questions and parts.

In the course of doing this, I fixed quite a few bugs, mainly in loading
an exam from JSON.

There are now events corresponding to each place a display or storage
method is called, so it should eventually be possible to replace these
calls with listeners in the corresponding display or storage code.

There's a bash script, tests/check-events.sh, which tries to find events
or signals that have not been documented or tested.