fix DOMcontentsubber and THTML
I had half-completed making THTML tokens not use jQuery!

The `value` attribute of a THTML token is now always an array of HTML
nodes.

DOMcontentsubber had a couple of iterations over the `childNodes` of an
element, which would get confused if you added or removed a node inside
the loop. It now uses `Array.from` to take a slice of the children.