when warnings would go outside the viewport, stick to the right
fixes #457

This adds a binding handler to the part warning display, to check if the
box would be outside the viewport. If so, its CSS is set to `right:
0.5em`, so that the right edge is just off the right edge of the
  viewport.

It's shifted down by the height of the parent part element, so that it
doesn't overlap with the input widget.

The box is set to always be the same width, so it doesn't get squished
when it's close to the edge of the viewport.