track deterministic functions
Now, when a `funcObj` object has the property `random === false`, i.e.
not true and not undefined, it is considered to be deterministic.

The function `Numbas.jme.isDeterministic` checks if an expression is
deterministic in its free variables.

This commit also changes `Numbas.jme.isRandom` so it checks for
random expressions substituted into strings.

Working towards #532