cancelTerms rule: the other part must contain a non-number
fixes #615

This isn't a great fix: the rule is quite a bit slower, and still
doesn't match some things that it should, in particular anything
involving division.

I feel like it should be possible to break this into smaller steps, but
the ;= operator isn't working how I expect.

The rule
    (?;=x * ?`*;a) + (?;=x * ?`*;b)
doesn't match
    x*y + z*x
I'd expect it to find the `x` that matches in both terms.
It seems to only ever match the first factor as the name `x`, and
doesn't try rearranging, even with commutative mode on.