the time remaining is always visible in sequence mode
fixes #652

The time remaining is shown in both the sidebar and the top nav.

On the sidebar, it has position: sticky, so it sticks to the bottom of
the screen if the sidebar is too long.

On the tap nav, it's pushed to the right, next to the hamburger button.