add rationaliseDenominators and reduceSurds rules, and largest_square_factor function
fixes #801

The rule 'rationaliseDenominators' finds factors of the form `/sqrt(a)` in the denominator of a
fraction, and replaces them by `sqrt(a)/a`.

The rule `reduceSurds` tries to pull out the largest square factor from
square roots, i.e. `sqrt(27)` → `3*sqrt(3)`. It looks for numbers with
square factors, and integer powers greater than 1 of anything else.

This also adds definitions of trunc, fract, ceil and floor for
rationals, producing integers or rationals, instead of decimals.