rejig part feedback messages
fixes #949

The feedback messages area no longer has max width or height.

Individual messages have max width and height based on the viewport
size.

The feedback icon is on the right, aligned with the final feedback icon.

The final score line has a horizontal line above it when the part has
been answered, separating it from the feedback messages.