when restoring 1_n_2 and m_n_2, set shuffleChoices = [0]
The `shuffleChoices` attribute isn't saved to the suspend data because
it's always the same for 1_n_2 and m_n_2 parts. But it needs to be
restored properly!

The change in ac4a93f8b13f811b299628d43cf39a15d983feb8 made the marking
algorithm use `shuffleChoices`, so we can't get away with it being
undefined any more.

fixes #961