Rewrite Numbas.jme.inferVariableTypes
fixes #963

The new plan is: for a function call, enumerate lists of arguments
by data type that each definition of the function can use.

For each argument, check for each possible corresponding data type
in the enumeration whether that type can be produced, and if so,
how it constrains the assignments of variable types.

Working through the arguments from left to right, the list of possible
signatures and corresponding variable type assignments is gradually
reduced.

Finally, return the first of the remaining assignments.