\simplify subs values into a tree, not a string
Until now, when substituting values into the expression in a LaTeX
\simplify command, Numbas.jme.subvars was used on the string, replacing
stuff between braces with the JME code representation of the result of
evaluating it.

This meant that `decimal` values were rendered as the function
`dec(...)`, which is rendered in TeX as `\textrm{...}`.

It's much better to just parse the expression, evaluate the bits where
substitutions are meant to happen to a single value, and replace the
subtree with that value.

This commit adds a function Numbas.jme.display.treeToLaTeX, which is
used by the \simplify command.