lots of improvements to the worksheet theme
The layout has changed: in landscape view, there is a sidebar containing
the generator controls, and a single exam is displayed to the right,
210mm wide to simulate an A4 page.

After you've generated the exams, there are lots of styling options:

* show the sheet ID?
* page breaks between questions?
* page margins, in mm
* background colour and text colour (in case you're just making a PDF,
  because printing full-colour page is a bad idea!)
* Text size, in pt. (1pt = 4/3 px)

While exams are being generated, progress is shown with a progress bar
and percentage display, and there's a cancel button.

The front page template is included, showing the number of questions.
time allowed, total marks available and introduction message.

`break-inside: avoid` is set on parts, so they don't get split across
several pages unless they have to.