math.rangeToList works in Decimal values
This adds math.rangeToDecimalList, and math.rangeToList just calls that
and converts to plain numbers.

Rounding error when adding two JS number values is unavoidable. By
working with fixed-precision Decimal values and then converting to plain
numbers, the result is as close as it can be to the true value.

fixes #989