update unicode mappings
Mainly, better treatment of punctuation characters.

There's now a dictionary of mappings for bracket characters, replacing
the static lists of left and right parentheses.

Connecting punctuation is valid instead of an underscore inside a name
token.

All dash characters are interpreted as equivalent to the minus sign `-`.

see numbas/unicode-math-normalization#5