THTML.value is an array
Previously, the `value` property of a THTML token was the `childNodes`
property of a div element. When the children are appended to another
element in the page, they're no longer in the `childNodes` NodeList.

This commit converts the NodeList to an Array, so that references to the
children are still kept even as they move around the DOM.