add an "introduction" item to the sidebar, above the questions
The "introduction" infopage shows the introduction text that was shown
on the front page.

This commit also fixes the question links in the sidebar so that they
can be pressed with the keyboard. (previously, the click listener was on
the parent <li> element, so pressing the link with the keyboard did
nothing)

fixes #990