strings produced by currency() are marked as latex; escape dollars in rendering latex strings
For example, `currency(4,"$","")` will produce `"$4"`. If the string is
not tagged as LaTeX, then when it's subsituted into LaTeX it's rendered
as `\textrm{$4}`, which produces a "no matching delimiter" error because
the dollar symbol is interpreted as opening mathmode.

Now that the output is tagged as LaTeX, the string is inserted into TeX
almost verbatim, just with the dollar sign escaped.