add pipe operator `|>`
See #894

The pipe operator is syntactic sugar to make code that processes a value
in several steps more readable.

`a |> b(...)` is equivalent to `b(a, ...)`.

The operator is left-associative, so `a |> b(1) |> c(2)` is equivalent to
`c(b(a, 1), 2)`

The languages Elm and Elixir also have a pipe operator `|>`. In Elm, the
piped value is used as the last argument, while in Elixir it's the first
argument. I've gone with using the first argument because, like Elixir,
JME functions aren't curried, so often the most important argument is
the first one, and later arguments are used to parameterise the
operation.