Link answer inputs to warning boxes with aria-errormessage and aria-invalid
fixes #1032

Parts have a `full_path` property, which is `path` with the question
number prepended. This is used to create unique IDs for elements
corresponding to that part.

CPT answer widgets now use template literals for the template string,
which just makes the code more readable.

The `part_aria_validity` knockout binding sets the `aria-invalid` and
`aria-errormessage` attributes on an element depending on the given
observable, which is always the part display's `hasWarnings` observable.
It uses the `part` binding to get a reference to the part display
object.

This also fixes an unclosed `<thead>` tag in the template for
`<matrix-input>`.

This also fixes a bug where mathematical expression parts, and CPTs with
mathematical expression answers, would throw an obscure error when the
answer is empty instead of "no answer submitted".
To do that, Numbas.jme.treesSame checks if either tree is null or
undefined.