lots of accessibility changes
Links are underlined and use browser default colours.

Fixed the alignment of the footer - when the `with-stylesheet` wrapper
was added, the CSS selector for it no longer matched.

Heading levels now make sense: the exam name is <h1> in sequence mode;
question title is <h2> and always shown; part headers are <h3>.

The lightbox has a big enough z-index to appear above the nav sidebar.

Event handlers to show the lightbox are attached to image elements
directly, instead of using jQuery deferred events on #questionDisplay.
This means that plain text in the question is no longer reported by
assistive tech as clickable.

The scoreFeedback element is initially marked with aria-busy, so it
isn't read out when the page loads. (Fixes #1036)

In menu mode, the whole row for a question is clickable. It now has an
aria-labelledby attribute so its label is the question's name.

The buttons at the bottom of the nav sidebar are in a <ul> tag instead
of a <div>, so they're correctly interpreted as a list.

The label on the hamburger toggle button is correctly localised.

The question content is now an <article> tag, and parts are <section>
tags.