use grid layout for the nav bars
fixes #1030

The nav bars for question navigation are now siblings of the <main> tag,
and all three are positioned with grid layout.

This commit removes the jasny-bootstrap library, which provided the
floating left nav bar.

On narrow screens when the main question navigation is toggled on, it
takes the full width of the screen and the main area is hidden. It's
shown again when you move to a question or info page.

This also fixes the max-width of the matrix sizing box, so that it
breaks to two lines when there isn't enough space.