CPT widgets take an ID parameter; m_n_x inputs are labelled
The widgets for custom part types take an `id` parameter, so that
elements in the expected answer can have different IDs to the student
input.

The radio buttons or checkboxes for a "match choices with answers" part
have the aria-labelledby attribute, referring to the headers for the
corresponding row and column.