Revert "Show scores in review mode if reviewShowScore is true but expected answers haven't been revealed."
This reverts commit a35a8559967128a9a7a27eaf583ccc88dfa5996a.

see #1042.

The reverted commit tried to make it so that scores could be shown
immediately after finishing the exam, but not expected answers.
However, this has made it impossible to *not* show scores immediately,
only once re-entering in review mode.

Until there are enough settings to express this, the safest thing is to
go with the old behaviour, where less information was revealed to
students.