some fixes for printed worksheet theme in Chrome
`overflow-y: hidden` on #everything stops Chrome showing a massive
vertical scrollbar.

Invisible elements are positioned absolutely way off to negative x and
y. When there are elements like this, Chrome seems to try to scale the page to
fit them in. This commit sets them to `display: none`, which seems to
fix the problem.