sidebar is 100vh; move the footer to the bottom of the main area; fix scrolling in Chrome
fixes #1051

The footer is now a child of the <main> tag, beneath the question or
infopage content.

The side nav is exactly 100vh high, and the question list expands to
fill the available space.

Chrome continues to do weird things with scrolling. The <html> tag
by default scrolls when the question content is longer than the
viewport, which must be a bug. This commit adds `html {overflow:hidden}`
to prevent this, and also hides scrollbars in print media just to be
sure.

<p> and <pre> tags no longer have a max width in print media.

The top nav takes `min-content` vertical space; before this, something
was making it take a lot of empty space.

The progress bar above the question list is much shorter.

The introduction is now displayed like a question, with the side or top
nav.

When the top nav is visible, pressing the backwards button from the
first question moves to the introduction, if there is one, and pressing
the forwards button from the introduction moves back to the last-viewed
question (which should be the first question, if you're just using the
arrows to move).