move display utility functions to display-util.js
Trying to reduce the amount of duplication of code between the default
theme and the worksheet theme.

This commit moves a few utility functions to display-util.js, which
display-base imports. They're under a new namespace Numbas.display_util,
but references under Numbas.display are kept for backwards compatibility
with existing custom themes.

fixes #1060