the entire sidebar scrolls on short viewports
A recent change madethe question list in the sidebar scroll independently, and
the other elements consume the rest of the viewport height.
When the viewport is very short, and the logo + exam title are very
tall, this could leave almost no space for the question list on short
viewports.

This commit adds a breakpoint of 45em to make the entire sidebar scroll,
so the control buttons aren't initially visible but you can at least use
the full height for showing the question list.

The exam title is shown at font-size: 1em, so should take less space.

The control buttons are display: flex, so the display options and pause
buttons will usually sit on the same line.

The timer has a slight box-shadow on its top edge so when it's rendered
over the question list it's visually separate.