adaptive marking: show the "part must be answered" warning in the warnings box
When a part depends on a previous part that must be answered, an error
message is shown in the part feedback area.
This commit makes it so that the message is also shown in the part's
warnings box next to the input widget, where students expect to see
warnings.

fixes #1068