set bracketed: true when rewriting unknown function names to multiplication
The test case is `z(x*y)^2`: we need to remember that `(x*y)` was
bracketed, so shouldn't be split up.

The problem is that the expression looks like this:

   ^
 ┌─┴──┐
 z    2
 │
 *
┌┴─┐
x  y

so because the algorithm works depth-first on the tree, the `z(x*y)` is
rewritten on its own, with no way of knowing it's the argument to an
exponentiation.

When the exponentiation is handled, the tree looks like

    ^
   ┌┴────┐
   *     2
┌──┴┐
z   *
   ┌┴─┐
   x  y

Until now, the algorithm would assuming that the factors on the left are
each bits of a single expanded name, and only extract the rightmost term,
`y`, to have the exponentiation applied to it:

    *
 ┌──┴──┐
 *     ^
┌┴─┐  ┌┴─┐
z  x  y  2

This would make sense if the original expression had been `zxy^2`, and
we're only allowing single-letter variable names.

The fix is to set `bracketed: true` on the `(x*y)` part after rewriting
the function call to a multiplication, and to make `extract_rightmost`
consider a bracketed expression as a single term.

fixes #1072