move question-nav inside the question <article>
Commit 6ea7cdd8 moved the main vertical scrollbar to the <article>
element for each question, so that the scrolling position for each
question would be remembered.
As a result, the question-nav bar at the bottom needs to be inside that
element.

This commit puts a copy of question-nav inside each question, so it's
now generated as part of the question XSLT.

This commit also moves the right-padding from the question-container to
the question element, so that the scrollbar doesn't overlap the content.