in print-only version of dropdowns, strip block-level tags
Block-level tags such as <p> are not allowed inside a <span>, so
when one is encountered, the <span> tag is automatically ended,
breaking the intended hierarchy.

When the choices for a dropdownlist in a gapfill have <p> tags in
their content, this led to the choice being rendered outside the
containing <span> tag for the part, so the knockout context is not right.

This changes the template to use a "no-paragraph" mode for the content
of dropdownlist options which strips out block-level tags.