Make jme.makeFast faster
My first go at `Numbas.jme.makeFast` worked with any number of arguments
and free variables, so it used `Array.map()` to work over them all.

For a fixed number of arguments for each operation and number of free
variables, the function for each operation can be written directly in
JS, without mapping over an array. This is a lot faster.

This commit changes `Numbas.jme.makeFast` to use fixed arity functions
when there are at most 5 free variables and 5 arguments to an operation.

The expression `sin(x/2.0 + y)` is about half as fast as the native JS
`Math.sin(x/2.0+y)`, which isn't bad!