add `imprecise(n)` for marking numbers with no precision
Commit 21006040d3d124fda0a0495b621232f4f2ca4aca made the suspend data
not wrap numbers in the `with_precision` function when both `precision`
and `precisionType` are `nothing`. I assumed this would do nothing, but
in fact it removes inferred precision information from the number
literal.

I still want to save some characters, so this commit adds a function
`imprecise(n)` which is equivalent to
`with_precision(n,nothing,nothing)`.