Numbas.jme.makeFast can cope with type casting.
It also replaces all integer constants with numbers, so that they don't
produce rationals when you divide them, which produce decimals when you
combine them with something else.
Decimals are very slow to construct, so it's good to avoid them.

It would be nice if Numbas.jme.inferTreeType had a way of picking
functions that avoid decimal and rational outputs when possible, but
this is a quick fix for now.