gap-fill parts are invalid if any gap is invalid.
Fixes #1093.

The score from any answered gaps is still counted, but the part is
marked as unanswered.

If this breaks any existing questions which expect the student not
to answer one or more gaps, they can be made to show "answered" by
changing the marking algorithm for the unanswered gaps to not mark an
empty answer as invalid.