simplifyFractions doesn't lose extra terms in the denominator
Looking at the rule ` ?;a / (?;b / ?;c) `, it should capture extra terms
and stick them on the end. This works for extra factors on top of the
fraction, but extra factors in the denominator were lost.

This is a bug, but I spent a couple of hours trying to get it to capture
them and failed. So this commit instead takes the easy way out and
explicitly captures the other terms.