only clear displayed part feedback messages if there really are none
In 8438c95e32a4b3b8a9e14343b53ef32d26a1c155, I didn't notice that there
are two conditions in the if statement: there are feedback messages, and
the question hasn't had answers revealed.

This splits the if statement into two steps so that the list of
displayed messages isn't cleared if there are messages, but the question
has been revealed.

see #1081