Numbas.display.typeset waits until there's a JME scope
There was a timing issue with reviewing an exam, where the Knockout
`latex` binding would try to run MathJax on an element while it wasn't
attached to the page document.

This changes `Numbas.display.typeset` so that it only sets MathJax going
when the element is attached to the page, and has an ancestor which
specifies a JME scope.

I also noticed that the info pages and the diagnostic feedback box
didn't have a JME scope set, so these are now set to the exam scope when
the display is initialised.