Fixes to do with scientific notation
fixes #1096

There are a few different fixes to the same problem here.

The regex for scientific notation now allows spaces before and after the
'e', since we saw students doing this. I don't think there's any chance
of ambiguity, because this notation style is used for a number on its
own, not in the JME parser.

The marking script for number entry parts uses `cleanedStudentAnswer`
when counting the significant figures in the student's answer, rather
than `studentAnswer`, so that any parsing differences in the scientific
notation format don't mess it up.