draw a button over all images which activates the lightbox
Fixes #1097

Instead of just registering a click handler on images, this wraps them
in a span and draws a button absolutely positioned in the top-right
corner of the image. Only the button activates the lightbox.

The button is only visible when the pointer hovers over the image, or it
has focus. It can still be tabbed to.