lambdas work!
Implementing lambdas in the parser was the right choice.

You can call an anonymous function immediately, e.g. `(x -> x+1)(1)`.

The builtin functions which took an expression and name(s) to bind now
also have forms that take lambdas. The implementation is much similar:
the old form is rewritten to use a lambda, and all the binding and
scoping rules are handled almost automatically.

This also adds a function `separate(condition, list)`, fixing #1088.