automatically submit parts when the input loses focus
with a view to #783

This adds a line to the default `inputEvents` handler for blur and focus
events, to submit a part when its input is blurred.

If focus moves to another input in that part, it's not submitted.

We tested this with several part types and with explore mode or adaptive
marking, but need to do more thorough testing with real questions, and
questions with unusual inputs such as interactive diagrams.

Need to check that the custom part type input widgets use these events.

The "submit part" button has changed to "save answer", really to
reassure students who want to proactively make sure their answer is
saved.