Focus events anywhere in a part affect the auto-submit logic.
I noticed that in a mathematical expression part, moving focus to the
MathJax preview caused the part to be auto-submitted, because the
MathJax element didn't use the part's inputEvents focus and blur
handlers.

This adds focusin and focusout event handlers to the part's container
element, so that a part is auto-submitted only when focus moves from the
part to something else. When navigating by keyboard, this is typically
when you move focus to the next part.