finishing up auto-submit?
This might finish #783

To get around the problem of new feedback causing the height of the part
to change, feedback messages are now in a <details> tag, which is closed
by default. It opens if you proactively submit the part, by clicking the
"save answer" button or by pressing Enter inside an input control.

The "answer all parts" button in the question nav now says "all answers saved" when
appropriate.
It now has a fixed width, so that the changing text doesn't cause the
width of the whole page to change.
The feedback icon in the question nav also has a fixed size.

The part score is now above the feedback messages, so it's in a
consistent place, and the border is always present so that it doesn't
change the height after the part is answered.

The "show feedback" line is only shown when the part has feedback
messages, or when waiting for pre-submit tasks.

In explore mode, the buttons to move to next parts are disabled when the
part is dirty, so that you can't move without auto-submitting a changed
answer.