worksheet theme waits for document ready before init
The base start-exam.js used to use jQuery's load function to wait for the document to be loaded before doing the init stuff.

The worksheet theme didn't do anything, so there could be a timing problem if the code ran before the page had finished loading.

This adds a function Numbas.util.document_ready, which uses the native document.readyState property to wait until the page has finished loading, and changes both versions of start-exam.js to use it.