Reorganised exam feedback options
This commit changes the .exam schema, rejigging how many of the feedback
settings work.

The aims are to:

* Make the feedback settings in the editor easier to understand
  (see numbas/editor#795)
* Allow a "completed but not revealed" mode, where students can see the
  questions and their answers, but not necessarily any feedback.

The settings `showActualMark`, `showTotalMark`, `showAnswerState` and
the new setting `showPartFeedbackMessages` are now strings, taking
values from ["always", "oncompletion", "inreview", "never"].

The settings `reviewShowScore` and `reviewShowFeedback` have been
removed, since their behaviour is described by the changed `showX` settings.

The `showResultsPage` setting has been removed and there's a new setting
`enterReviewModeImmediately` which determines what happens when you end
the exam.

The settings `reviewShowExpectedAnswer` and `reviewShowAdvice` are now
`revealExpectedAnswers` and `revealAdvice`, with values in ["inreview",
"never"].

These settings give more fine-grained control over when feedback is
revealed: you might want to show students expected answers as soon as
the exam is finished, but not show how their answers were marked until a
later date.

We imagine that "enterReviewModeImmediately" will be false for summative
assessments, and the LTI tool will let students open attempts in review
mode once the deadline has passed.

There's an exam data migration which automatically upgrades old exams to
the new schema.

We still need to add unit tests for this behaviour. I'm not sure how
much of it can be tested, since a lot of the logic is in the display
code.