jme.find_valid_assignments: immediately return the first good assignment
fixes #1109

In Numbas.jme.find_valid_assignments, when trying to find an assignment
to satisfy a particular function or operator, we make a list of all
possible assignments, then filter out ones that don't work with each of
the arguments.

The old code was building up a big list of valid assignments and then
returning the first one.

This wasted a lot of effort, especially when there are lots of options
and a very deep expression tree.

This chnages the routine to return immediately as soon as it finds an
assignment that works, which would be the same assignment it would have
returned with the old code.