all colour literals are now in _variables!
There are now no colour literals outside _variables.css.

I have a clearer idea of which colours need to be defined.
I have a test page at
https://automatic-colour-scheme.think.somethingorotherwhatever.com/
which tries to compute shades of colours based on just a few input
colours. It relies a lot on the color-mix() CSS function which is
currently at 89% compatibility. We could either wait for bigger uptake,
or implement the mixing in JavaScript, which wouldn't be hard but would
make the stylesheet a lot less tidy.

This commit also rewrites the modals to use the <dialog> element's
native behaviour instead of bootstrap's modal script.