number_to_decimal takes the min of 15 and the input number's precision
If a `number` value has been rounded to less than 15 decimal places,
which we say is the maximum reliable precision for a float, then round
the decimal value to that many decimal places.

The use of precision was added by
5575f10afcf0ae6747ba76c3d0c8991b4378f0c2, which tried to account for the
size of numbers used as the expected answers for number entry parts, but
that made a sign error. This commit fixes the sign error, so those
values are consistent with how precision is represented in `number`
values made elsewhere.