lots of jiggery-pokery to fix scrolling
I noticed that the bottom of the question nav bar was cut off on a
question in a diagnostic mode exam.

I've had another go at setting max-height on elements leading down to
the question content so that it has the right size and scrolls all the
way to the bottom.

I've checked in Firefox, Chrome and Safari. In Safari 17 there's still a
layout bug with the results screen - it seems too wide for the page.