separate the parts of feedback messages about credit
In order to comply with the new feedback settings, it should be possible
to show part feedback messages but not the student's current score.

To do this, the part of the message about the credit (e.g. "you were
awarded 1 mark") is now in a separate attribute, `credit_message`, which
is only shown if `showActualMark` is true.