show expected answers when the question is revealed
After the recent changes to feedback, expected answers were only shown
when the exam is in the `revealed` state.

When you do "Reveal answers" on a single question, the exam is not
`revealed`, only the question.

This changes the logic for displaying expected answers to use the
question's `revealed` state instead of the exam's.

This still leaves a weird behaviour where if "Allow revealing answers to
a single question" is turned on, but all feedback is set to "only in
review mode" and "enter review mode immediately on ending the exam" is
turned off, then you can see your scores for a question by revealing
answers, but any questions you haven't revealed when the exam ends don't
show feedback.

Should "allow revealing answers to a single question" only be available
if "enter review mode immediately on ending the exam" is turned off, and
"show expected answers" isn't "never"?

fixes #1115