add another rule for removing brackets involving subtraction
This rule exists to cover a really fiddly case which emerges because of how unary minus is handled inside a product.

When looking for additive terms in the expression `-a - 2*b`, the second term is rewritten so that the unary minus is at the bottom of the tree: `(-2) * b` instead of `-(2*b)`.

This means that the expression ends up looking like `-a + ((-2)*b)`, so it doesn't match the rule `(-?;b - ?;c)`, because the unary minus is too deep.

I've spent a while trying to get the matcher to look deeper for unary minus, but that led to a lot of other things breaking.

In the end, I've added another version of the rule which deals with a `+` inside the brackets instead of a `-`.