main content tries to be as wide as possible
On wide screens, the main container had a max-width set but not a width,
so it would be narrow when the question content is narrow. Feedback
messages could it make it wider, which leads to things moving around
unexpectedly.

This changes the container to be as wide as it can, up to the limit of
1170px or the viewport width, whichever is less.