some improvements to screen reader part feedback
see #1122

The "Feedback for part <name>" text is in the summary of the feedback
box, so it's read out before the "Show feedback" text.

I've tried to fix the href on the skip-link, but pressing the link with
VoiceOver turned on still doesn't seem to do anything. Pressing the link
without VoiceOver on does move focus to the start of the question
content.

When I added the pointerdown/up handlers to parts, I didn't notice there
was already an event handler for focusin/out, so they stopped working.
I've collected them into PartDisplay.event_handlers.