Offline analysis: ability to review attempts
This adds the ability to review attempts to the offline analysis tool.

In the list of attempt data, there's a button for each attempt labelled
"Review this attempt". It loads the exam in an iframe and creates a
SCORM API to review the attempt data.

The downloaded data only includes the most recent version of the suspend
data object, so we can't offer a timeline view.

fixes #1105